Support arbitrary shapes for metrics, not just accuracy
Our project has some model-specific metrics we want to collect, so I'm
generalizing to support arbitrary metric shapes collected at each
sample.

The only interpretation we apply to the metrics is summing the
metrics pointwise over the dataset -- then a metric of the same shape
is returned. So if you want to calculate means, add a count metric
that's always one.

Removed error_rate and StopWhenAccurate for now because they don't
make sense without explicit accuracy.